source/site/requirements.txt
dash==2.15.0
face_detection==1.0.5
fastapi==0.110.0
matplotlib==3.8.2
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
pandas==2.2.1
Pillow==10.1.0
Pillow==10.2.0
plotly==5.19.0
pyttsx3==2.90
scipy==1.12.0
torch==2.1.2+cu121
torchvision==0.16.2+cu121
